Divided and updated tests.
