+++ Tests/test_PDB_Disordered.py
@@ -10,6 +10,9 @@
 import os
 import tempfile
 import unittest
+import warnings
+
+import numpy as np
 
 from Bio.PDB import PDBParser, PDBIO
 
@@ -17,16 +20,18 @@ from Bio.PDB import PDBParser, PDBIO
 class TestDisordered(unittest.TestCase):
     """Tests for operations on DisorderedEntities."""
 
+    @classmethod
+    def setUpClass(cls):
+        cls.parser = parser = PDBParser(QUIET=1)
+        cls.structure = parser.get_structure("x", "PDB/disordered.pdb")
+
     def unpack_all_atoms(self, structure):
         """Return a list of all atoms in the structure."""
         return [a for r in structure.get_residues() for a in r.get_unpacked_list()]
 
     def test_copy_disordered_atom(self):
         """Copies disordered atoms and all their children."""
-        parser = PDBParser(QUIET=1)
-        s = parser.get_structure("x", "PDB/disordered.pdb")
-
-        resi27 = s[0]["A"][27]
+        resi27 = self.structure[0]["A"][27]
         resi27_copy = resi27.copy()
 
         self.assertNotEqual(id(resi27), id(resi27_copy))  # did we really copy
@@ -40,9 +45,7 @@ class TestDisordered(unittest.TestCase):
 
     def test_copy_entire_chain(self):
         """Copy propagates throughout SMCRA object."""
-        parser = PDBParser(QUIET=1)
-        s = parser.get_structure("x", "PDB/disordered.pdb")
-
+        s = self.structure
         s_copy = s.copy()
 
         self.assertNotEqual(id(s), id(s_copy))  # did we really copy
@@ -58,9 +61,7 @@ class TestDisordered(unittest.TestCase):
         """Transform propagates through disordered atoms."""
         # This test relates to issue #455 where applying a transformation
         # to a copied structure did not work for disordered atoms.
-        parser = PDBParser(QUIET=1)
-        s = parser.get_structure("x", "PDB/disordered.pdb")
-
+        s = self.structure
         s_copy = s.copy()
 
         mtx = ((1, 0, 0), (0, 1, 0), (0, 0, 1))
@@ -76,10 +77,9 @@ class TestDisordered(unittest.TestCase):
 
     def test_copy_and_write_disordered(self):
         """Extract, save, and parse again disordered atoms."""
-        parser = PDBParser(QUIET=1)
         writer = PDBIO()
 
-        s = parser.get_structure("x", "PDB/disordered.pdb")
+        s = self.structure
 
         # Extract the chain object
         chain = s[0]["A"]
@@ -92,7 +92,7 @@ class TestDisordered(unittest.TestCase):
             writer.save(filename)
 
             # Parse again
-            s2 = parser.get_structure("x_copy", filename)
+            s2 = self.parser.get_structure("x_copy", filename)
 
             # Do we have the same stuff?
             atoms1 = self.unpack_all_atoms(s)
@@ -103,6 +103,37 @@ class TestDisordered(unittest.TestCase):
 
         finally:
             os.remove(filename)
+
+    # DisorderedAtom.center_of_mass
+    def test_structure_w_disordered_com(self):
+        """Calculate center of mass of structure including DisorderedAtoms."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            s = self.parser.get_structure("b", "PDB/disordered.pdb")
+
+        com = s.center_of_mass()
+
+        self.assertTrue(np.allclose(com, [54.545, 19.868, 31.212], atol=1e-3))
+
+    def test_disordered_cog(self):
+        """Calculate DisorderedAtom center of geometry."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            s = self.parser.get_structure("b", "PDB/disordered.pdb")
+
+        arg27 = s[0]["A"][27]
+
+        # Detach all children but NH1
+        for atom in list(arg27):
+            if atom.name != "NH1":
+                arg27.detach_child(atom.name)
+
+        res_cog = arg27.center_of_mass()
+        self.assertTrue(np.allclose(res_cog, [59.555, 21.033, 25.954], atol=1e-3))
+
+        # Now compare to DisorderedAtom.center_of_mass
+        da_cog = arg27["NH1"].center_of_mass()
+        self.assertTrue(np.allclose(res_cog, da_cog, atol=1e-3))
 
 
 if __name__ == "__main__":
